Revert "refactor : Url 엔티티 관련 수정"
This reverts commit 8be734e1635e14feb00e2010da82f3247b68e700.
