+++ src/main/java/com/programmers/springbooturlshortener/domain/Url.java
@@ -1,14 +1,10 @@
-package com.programmers.springbooturlshortener.domain.url;
+package com.programmers.springbooturlshortener.domain;
+
+import javax.persistence.*;
 
 import lombok.AccessLevel;
 import lombok.Builder;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.lang.Nullable;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
 @Entity
 @Table(
@@ -21,22 +17,19 @@ import javax.validation.constraints.NotNull;
         }
 )
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
 public class Url {
 
     @Id
     @GeneratedValue
     private Long id;
 
-    @NotNull
+    @Enumerated(EnumType.STRING)
     @Column(name = "algorithm", length = 20)
-    private String algorithm;
+    private Algorithm algorithm;
 
-    @NotBlank
-    @Column(name = "origin_url", nullable = false, length = 2000)
+    @Column(name = "origin_url", length = 2000)
     private String originUrl;
 
-    @Nullable
     @Column(name = "short_url", length = 7)
     private String shortUrl;
 
@@ -44,17 +37,14 @@ public class Url {
     private Long requestCount;
 
     @Builder
-    public Url(String algorithm, String originUrl, Long requestCount) {
+    public Url(Algorithm algorithm, String originUrl, String shortUrl, Long requestCount) {
         this.algorithm = algorithm;
         this.originUrl = originUrl;
+        this.shortUrl = shortUrl;
         this.requestCount = requestCount;
     }
 
     public void increaseRequestCount() {
         this.requestCount++;
-    }
-
-    public void setShortUrl(String shortUrl) {
-        this.shortUrl = shortUrl;
     }
 }
